alumnos enlazado, css de profes curso hecho

+++ profesor/alumnos/funciones.php
@@ -1,30 +1,30 @@
 <?php
-
     function mostrarImg($conn){
         $idProfe = $_SESSION['idProfe'];
-
         $queryName = 'SELECT img, tipus FROM profesores WHERE id = ' . $idProfe;
-
         $r = mysqli_query($conn, $queryName);
 
-        // Asegúrate de que haya resultados
         if ($r && mysqli_num_rows($r) > 0) {
             $fila = mysqli_fetch_assoc($r);
-
-            // Mostrar la imagen a través de un archivo PHP
             echo "<img src='data:" . $fila['tipus'] . ";base64," . base64_encode($fila['img']) . "' >";
         }
     }
     
     function updateAlumno($conn) {
         $alumno_id = intval($_POST['alumno_id']);
-        $usernameA = $_POST["usernameM"];
-        $nombreA = $_POST["nombreM"];
-        $apellidoA = $_POST["apellidoM"];
-        $contrasenyaA = $_POST["contrasenyaM"];
-    
-        $update = "UPDATE alumnos SET username = '$usernameA', pass = '$contrasenyaA', name = '$nombreA', last_name = '$apellidoA' WHERE id = '$alumno_id'";
-        mysqli_query($conn, $update);
+        if(!empty($alumno_id)){
+            $usernameA = $_POST["usernameM"];
+            $nombreA = $_POST["nombreM"];
+            $apellidoA = $_POST["apellidoM"];
+            $contrasenyaA = $_POST["contrasenyaM"];
+        
+            $update = "UPDATE alumnos SET username = '$usernameA', pass = '$contrasenyaA', name = '$nombreA', last_name = '$apellidoA' WHERE id = '$alumno_id'";
+            mysqli_query($conn, $update);
+        }else{
+            echo '<alert>Selecciona un Alumno</alert>';
+        }
+        header('Location: alumnos.php');
+        exit();
     }
 
     function crearAlumno($conn) {
@@ -40,20 +40,22 @@
             return;
         } else {
             $insert = "INSERT INTO alumnos (username, pass, name, last_name, curso_id) VALUES ('$usernameA', '$contrasenyaA', '$nombreA', '$apellidoA', '$curso')";
-    
             mysqli_query($conn, $insert);
+            header('Location: alumnos.php');
+            exit();
         }
-    
     }
 
     function eliminarAlumno($conn) {
         $alumno_id = intval($_POST["alumno_id"]);
-
-        $delete1 = "DELETE FROM alumnos WHERE id = $alumno_id";
-        $delete2 = "DELETE FROM scroes WHERE student_id = $alumno_id";
-
-        mysqli_query($conn, $delete2);
-        mysqli_query($conn, $delete1);
-        
+        if(!empty($alumno_id)){
+            $alumno_id = intval($_POST["alumno_id"]);
+            $delete1 = "DELETE FROM alumnos WHERE id = $alumno_id";
+            mysqli_query($conn, $delete1);
+        }else{
+            echo '<alert>Selecciona un Alumno</alert>';
+        }
+        header('Location: alumnos.php');
+        exit();
     }
 ?>
